backend/requirements.txt
fastapi==0.115.2
uvicorn[standard]==0.32.0
SQLAlchemy==2.0.49
psycopg2-binary>=2.9.9
pydantic-settings==2.5.2
python-jose[cryptography]==3.4.0
bcrypt==4.1.3
alembic==1.13.3
email-validator==2.3.0
celery==5.4.0
redis==5.1.1
python-multipart>=0.0.9
motor>=3.3.0
pymongo>=4.6.0
certifi>=2024.2.2
